Added Transactions Fragment
