+++ app/build.gradle
@@ -25,7 +25,8 @@ dependencies {
     implementation 'com.android.support.constraint:constraint-layout:1.1.3'
 
     //Web3j
-    compile ('org.web3j:core:3.3.1-android')
+    implementation 'com.android.support:support-v4:26.1.0'
+    compile('org.web3j:core:3.3.1-android')
 
     testImplementation 'junit:junit:4.12'
     androidTestImplementation 'com.android.support.test:runner:1.0.2'
