+++ app/src/main/res/values/strings.xml
@@ -3,4 +3,7 @@
     <string name="password_edttxt">Input a strong password</string>
     <string name="password_lbl">Don\'t lose this password</string>
     <string name="submit">Submit</string>
+
+    <!-- TODO: Remove or change this placeholder text -->
+    <string name="hello_blank_fragment">Hello blank fragment</string>
 </resources>
